+++ tests/mocker/v1/index.js
@@ -1,0 +1,5 @@
+require('./announcements');
+require('./contacts');
+require('./events');
+require('./faqs');
+require('./resources');