+++ CHANGELOG.md
@@ -5,6 +5,15 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+### Added
+- Add tests for all resources.
+
+### Changed
+- Use `axios` for all network requests.
+- Distribute package as UMD webpack bundle.
+
+### Fixed
+- Fix incorrect events sort order.
 
 ## [1.3.1] - 2018-02-11
 ### Fixed